TelegramBot/requirements.txt
aiogram
openai==0.28
python-dotenv
numpy==1.24.3
pydantic==1.8.1
clyent==1.2.1
nbformat==5.4.0